feat(chart): renamed files
