+++ packages/chart/src/components/tooltip-content/tooltip-icon/index.module.css
