+++ packages/chart/src/components/tooltip-content/tooltip-icon/index.tsx
@@ -4,8 +4,8 @@ import cn from 'classnames';
 import { LegendProps } from '../../../types/legend.types';
 import { SeriaProps } from '../../../types/seria.types';
 
-import { CircleIcon } from '../../../icons/Circle';
-import { icons } from '../../Legends';
+import { CircleIcon } from '../../../icons/circle-icon';
+import { icons } from '../../legends';
 
 import styles from './index.module.css';
