+++ packages/chart/src/components/tooltip-content/tooltip-value/index.module.css
