+++ packages/chart/src/hoc/customized-hoc.jsx
