+++ packages/chart/src/hooks/use-settings/utils/set-gradient-charts.ts
