+++ packages/chart/src/hooks/use-settings/utils/sort-by-index.ts
