+++ packages/chart/src/icons/circle-icon.tsx
