+++ packages/chart/src/icons/filled-circle-icon.tsx
