+++ packages/chart/src/icons/point-icon.tsx
