+++ packages/chart/src/icons/stroke-circle-icon.tsx
